+++ src/OpenTelemetry.Instrumentation.GrpcNetClient/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* **BREAKING**: Update to version 1.41.0 of the Semantic Conventions.
+* **BREAKING**: Update to version 1.42.0 of the Semantic Conventions.
   ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338),
   [#4508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4508))